DRY up SoaMailer classes
